refactor : 디렉터리 구조 수정
