+++ app/src/main/AndroidManifest.xml
@@ -15,12 +15,12 @@
         android:theme="@style/Theme.GOSOPTAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".signUp.SignUpActivity"
+            android:name=".presentation.signUp.SignUpActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize|stateAlwaysHidden" />
         <activity
-            android:name=".sginIn.SignInActivity"
+            android:name=".presentation.sginIn.SignInActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -29,7 +29,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".home.FragmentManageActivity"
+            android:name=".presentation.main.FragmentManageActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize|stateAlwaysHidden">
