+++ app/src/main/java/org/android/go/sopt/data/remote/dto/SignUpDto.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data.remote.model
+package org.android.go.sopt.data.remote.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
