+++ app/src/main/java/org/android/go/sopt/data/remote/service/SignInService.kt
@@ -1,7 +1,7 @@
 package org.android.go.sopt.data.remote.service
 
-import org.android.go.sopt.data.remote.model.RequestSignInDto
-import org.android.go.sopt.data.remote.model.ResponseSignInDto
+import org.android.go.sopt.data.remote.dto.RequestSignInDto
+import org.android.go.sopt.data.remote.dto.ResponseSignInDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
