+++ app/src/main/java/org/android/go/sopt/data/remote/service/SignUpService.kt
@@ -1,7 +1,7 @@
 package org.android.go.sopt.data.remote.service
 
-import org.android.go.sopt.data.remote.model.RequestSignUpDto
-import org.android.go.sopt.data.remote.model.ResponseSignUpDto
+import org.android.go.sopt.data.remote.dto.RequestSignUpDto
+import org.android.go.sopt.data.remote.dto.ResponseSignUpDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
