+++ app/src/main/java/org/android/go/sopt/presentation/main/gallery/GalleryAdapter.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.home.adapter
+package org.android.go.sopt.presentation.main.gallery
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
