+++ app/src/main/java/org/android/go/sopt/presentation/viewmodel/GalleryViewModel.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.home.viewmodel
+package org.android.go.sopt.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
 import org.android.go.sopt.R
