+++ app/src/main/java/org/android/go/sopt/presentation/viewmodel/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.home.viewmodel
+package org.android.go.sopt.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
 import org.android.go.sopt.data.remote.ServicePool
