+++ app/src/main/res/layout/fragment_gallery.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.GalleryFragment">
+    tools:context=".presentation.main.gallery.GalleryFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/pager_gallery"
